Change Types to Byte, Refactor Encoding to create even less additional objects

+++ src/domain/criptography/BinaryData.java
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by ignas on 15.11.15.
  */
 public interface BinaryData {
-    List<Integer> getBytes();
+    List<Byte> getBytes();
     BinaryStream getStream();
     boolean isValid();
 }
